Finish layout for the show page
"who to follow" for the logged in user's show page, and "followed
by" for other users' show page. User stats in the right side tab.
User method to return "users to follow" in the correct order, with
the attributes needed to be rendered. Controller returns the correct
users list depending on the show page being rendered.